FacialBurns {lavaan}	R Documentation

Dataset, originally used for illustrating the InformativeTesting function.

Description

A dataset from the Dutch burn center (http://www.adbc.nl). The data were used to examine psychosocial functioning in patients with facial burn wounds. Psychosocial functioning was measured by Anxiety and depression symptoms (HADS), and self-esteem (Rosenberg's self-esteem scale).

Usage

data(FacialBurns)

Format

A data frame of 77 observations of 6 variables.

Selfesteem

Rosenberg's self-esteem scale

HADS

Anxiety and depression scale

Age

Age measured in years, control variable

TBSA

Total Burned Surface Area

RUM

Rumination, control variable

Sex

Gender, grouping variable

Examples

head(FacialBurns)

[Package lavaan version 0.6-21 Index]