lavaan-deprecated {lavaan}	R Documentation

Deprecated functions in lavaan package

Description

Deprecated functions in lavaan.

Details

The functions in the first column hereafter are deprecated and calls should be replaced by the calls of the corresponding function in the second column.

# bootstrapLavaan 	 lavBootstrap

char2num 	 lav_char2num

cor2cov 	 lav_cor2cov

getCov 	 lav_getcov

inspectSampleCov 	 lavInspectSampleCov

# lavaanNames 	 lavNames

mplus2lavaan.modelSyntax 	 lav_mplus_syntax_model

mplus2lavaan 	 lav_Mplus_lavaan

parameterestimates 	 lavParameterEstimates

# simulateData 	 lavSimulateData

[Package lavaan version 0.6-21 Index]