lavaanList-class {lavaan}	R Documentation

Class For Representing A List of (Fitted) Latent Variable Models

Description

The lavaanList class represents a collection of (fitted) latent variable models, for a (potentially large) number of datasets. It contains information about the model (which is always the same), and for every dataset a set of (user-specified) slots from a regular lavaan object.

Objects from the Class

Objects can be created via the cfaList, semList, or lavaanList functions.

Slots

call:

The function call as returned by match.call().

Options:

Named list of options that were provided by the user, or filled-in automatically.

ParTable:

Named list describing the model parameters. Can be coerced to a data.frame. In the documentation, this is called the ‘parameter table’.

pta:

Named list containing parameter table attributes.

Data:

Object of internal class "Data": information about the data.

Model:

Object of internal class "Model": the internal (matrix) representation of the model

meta:

List containing additional flags. For internal use only.

timingList:

List. Timing slot per dataset.

ParTableList:

List. ParTable slot per dataset.

DataList:

List. Data slot per dataset.

SampleStatsList:

List. SampleStats slot per dataset.

CacheList:

List. Cache slot per dataset.

vcovList:

List. vcov slot per dataset.

testList:

List. test slot per dataset.

optimList:

List. optim slot per dataset.

impliedList:

List. implied slot per dataset.

h1List:

List. h1 slot per dataset.

loglikList:

List. loglik slot per dataset.

baselineList:

List. baseline slot per dataset.

funList:

List. fun slot per dataset.

external:

List. Empty slot to be used by add-on packages.

Methods

coef

signature(object = "lavaanList", type = "free"): Returns the estimates of the parameters in the model as the columns in a matrix; each column corresponds to a different dataset. If type="free", only the free parameters are returned. If type="user", all parameters listed in the parameter table are returned, including constrained and fixed parameters.

summary

signature(object = "lavaanList", header = TRUE, estimates = TRUE, nd = 3L): Print a summary of the collection of fitted models. If header = TRUE, the header section is printed. If estimates = TRUE, print the parameter estimates section. The argument nd determines the number of digits after the decimal point to be printed (currently only in the parameter estimates section.) Nothing is returned (use parameterEstimates or another extractor function to extract information from this object).

See Also

cfaList, semList, lavaanList

[Package lavaan version 0.6-8 Index]