varTable {lavaan}	R Documentation

Variable Table

Description

Summary information about the variables included in either a data.frame, or a fitted lavaan object.

Usage

varTable(object, ov.names = names(object), ov.names.x = NULL, 
    ordered = NULL, factor = NULL, as.data.frame. = TRUE)

Arguments

object
Either a data.frame, or an object of class lavaan.

ov.names
Only used if object is a data.frame. A character vector containing the variables that need to be summarized.

ov.names.x
Only used if object is a data.frame. A character vector containing additional variables that need to be summarized.

ordered
Character vector. Which variables should be treated as ordered factors

factor
Character vector. Which variables should be treated as (unordered) factors?

as.data.frame.
If TRUE, return the list as a data.frame.

Value

A list or data.frame containing summary information about variables in a data.frame. If object is a fitted lavaan object, it displays the summary information about the observed variables that are included in the model. The summary information includes variable type (numeric, ordered, ...), the number of non-missing values, the mean and variance for numeric variables, the number of levels of ordered variables, and the labels for ordered variables.

Examples

HS.model <- ' visual  =~ x1 + x2 + x3
              textual =~ x4 + x5 + x6
              speed   =~ x7 + x8 + x9 '

fit <- cfa(HS.model, data=HolzingerSwineford1939)
varTable(fit)

[Package lavaan version 0.6-8 Index]