lavNames {lavaan}	R Documentation

lavaan Names

Description

Extract variables names from a fitted lavaan object.

Usage

lavNames(object, type = "ov", ...)

Arguments

object
An object of class lavaan.

type
Character. The type of variables whose names should be extracted. See details for a complete list.

...
Additional selection variables. For example "group = 2L" (in a multiple-group analysis) only considers the variables included in the model for the second group.

Details

The order of the variable names, as returned by lavNames determines the order in which the variables are listed in the parameter table, and therefore also in the summary output.

The following variable types are available:

"ov": observed variables

"ov.x": (pure) exogenous observed variables (no mediators)

"ov.nox": non-exogenous observed variables

"ov.model": modelled observed variables (joint vs conditional)

"ov.y": (pure) endogenous variables (dependent only) (no mediators)

"ov.num": numeric observed variables

"ov.ord": ordinal observed variables

"ov.ind": observed indicators of latent variables

"ov.orphan": lonely observed variables (only intercepts/variancesappear in the model syntax)

"ov.interaction": interaction terms (defined by the colon operator)

"th": threshold names ordinal variables only

"th.mean": threshold names ordinal + numeric variables (if any)

"lv": latent variables

"lv.regular": latent variables (defined by =~ only)

"lv.formative": latent variables (defined by <~ only)

"lv.x": (pure) exogenous variables

"lv.y": (pure) endogenous variables

"lv.nox": non-exogenous latent variables

"lv.nonnormal": latent variables with non-normal indicators

"lv.interaction": interaction terms at the latent level

"eqs.y": variables that appear as dependent variables in a regression formula (but not indicators of latent variables)

"eqs.x": variables that appear as independent variables in a regression formula

See Also

lavaanify, parTable

Examples

HS.model <- ' visual  =~ x1 + x2 + x3
              textual =~ x4 + x5 + x6
              speed   =~ x7 + x8 + x9 '

fit <- cfa(HS.model, data=HolzingerSwineford1939)
lavNames(fit, "ov")

[Package lavaan version 0.6-8 Index]