lav_model {lavaan}	R Documentation

lavaan model functions

Description

Utility functions related to internal model representation (lavmodel)

Usage

# set/get free parameters
lav_model_set_parameters(lavmodel, x = NULL)
lav_model_get_parameters(lavmodel, GLIST = NULL, type = "free",
                         extra = TRUE)

# compute model-implied statistics
lav_model_implied(lavmodel, GLIST = NULL)

# compute standard errors
lav_model_vcov_se(lavmodel, lavpartable, VCOV = NULL, BOOT = NULL)

Arguments

lavmodel
An internal representation of a lavaan model.

x
Numeric.

A vector containing the values of all the free model parameters.

GLIST
List. A list of model matrices, similar to the output of lavInspect(object, "est").

type
Character string. If "free", only return the free model parameters. If "user", return all the parameters (free and fixed) as they appear in the user-specified parameter table.

extra
Logical. If TRUE, also include values for rows in the parameter table where the operator is one of ":=", "==", "<" or ">".

lavpartable
A parameter table.

VCOV
Numeric matrix containing an estimate of the variance covariance matrix of the free model parameters.

BOOT
Numeric matrix containing the bootstrap based parameter estimates (in the columns) for each bootstrap sample (in the rows).

Examples

HS.model <- ' visual  =~ x1 + x2 + x3
              textual =~ x4 + x5 + x6
              speed   =~ x7 + x8 + x9 '

fit <- cfa(HS.model, data=HolzingerSwineford1939)
lavmodel <- fit@Model

est <- lav_model_get_parameters(lavmodel)
est

[Package lavaan version 0.6-8 Index]