lavaan-deprecated {lavaan}	R Documentation

Deprecated Functions in the lavaan package

Description

These functions have been renamed, or have been removed. They are still included in this version for convenience, but they may be eventually removed.

Usage

vech(S, diagonal = TRUE)
vechr(S, diagonal = TRUE)
vechu(S, diagonal = TRUE)
vechru(S, diagonal = TRUE)
vech.reverse(x, diagonal = TRUE)
vechru.reverse(x, diagonal = TRUE)
vechr.reverse(x, diagonal = TRUE)
vechu.reverse(x, diagonal = TRUE)
lower2full(x, diagonal = TRUE)
upper2full(x, diagonal = TRUE)
duplicationMatrix(n = 1L)
commutationMatrix(m = 1L, n = 1L)
sqrtSymmetricMatrix(S)

Arguments

S
A symmetric matrix.

x
A numeric vector containing the lower triangular or upper triangular elements of a symmetric matrix, possibly including the diagonal elements.

diagonal
Logical. If TRUE, the diagonal is included. If FALSE, the diagonal is not included.

n
Integer. Dimension of the symmetric matrix, or column dimension of a non-square matrix.

m
Integer. Row dimension of a matrix.

Details

The vech function has been renamed lav_matrix_vech.

The vech.reverse function has been renamed lav_matrix_vech_reverse.

The lower2full function has been renamed lav_matrix_lower2full.

The duplicationMatrix function has been renamed lav_matrix_duplication.

The commutationMatrix function has been renamed lav_matrix_commutation.

The sqrtSymmetricMatrix function has been renamed lav_matrix_symmetric_sqrt.

[Package lavaan version 0.6-2 Index]